Subject: Fixing the dynamic library linking issue seen with dmapp_locks Description: GA used a lock descriptor array in its global space. With dynamic linking, DMAPP was not able to grab this region and register it during DMAPP_Init. This issue is encountered only when dynamic linking is used.
This mod introduces a fix for this issue. We now force GA to use
dmapp_sheap_malloc() to ensure that the lock descriptor array
is in the symmetric space.

With dmapp_sheap_malloc(), there is a risk of running of symmetric
memory space. In such cases, users need to adjust the value of
XT_SYMMETRIC_HEAP_SIZE to get past this issue. We will continue
exploring a solution that does not rely on dmapp_sheap_malloc().

svn path=/ga/trunk/; revision=1357